requirements.txt
pandas
beautifulsoup4
openpyxl
chardet
python-docx
requests
deep_translator
tenacity
dotenv